+++ packages/api/models/granules.js
@@ -1,12 +1,10 @@
 'use strict';
 
 const cloneDeep = require('lodash/cloneDeep');
-const get = require('lodash/get');
 const isArray = require('lodash/isArray');
 const isString = require('lodash/isString');
 const partial = require('lodash/partial');
 const path = require('path');
-const pMap = require('p-map');
 
 const awsClients = require('@cumulus/aws-client/services');
 const Lambda = require('@cumulus/aws-client/Lambda');
@@ -175,6 +173,9 @@ class Granule extends Manager {
     await cmr.deleteGranule(metadata.title, granule.collectionId);
   }
 
+  /*
+  * DEPRECATED: This has moved to /lib/granule-rmove-from-cmr.js
+  */
   async removeGranuleFromCmrByGranule(granule) {
     await this._removeGranuleFromCmr(granule);
     return this.update({ granuleId: granule.granuleId }, { published: false }, ['cmrLink']);
@@ -585,16 +586,6 @@ class Granule extends Manager {
    * @private
    */
   async _deleteRecord(granule) {
-    // Delete granule files
-    await pMap(
-      get(granule, 'files', []),
-      (file) => {
-        const bucket = this.fileUtils.getBucket(file);
-        const key = this.fileUtils.getKey(file);
-        return s3Utils.deleteS3Object(bucket, key);
-      }
-    );
-
     return super.delete({ granuleId: granule.granuleId });
   }