+++ packages/s3-credentials-endpoint/webpack.config.js
@@ -1,8 +1,22 @@
 'use strict';
 
 const path = require('path');
+const { IgnorePlugin } = require('webpack');
 // path to module root
 const root = path.resolve(__dirname);
+
+const ignoredPackages = [
+  'mssql',
+  'mssql/lib/base',
+  'mssql/package.json',
+  'mysql',
+  'mysql2',
+  'oracledb',
+  'pg-native',
+  'pg-query-stream',
+  'sqlite3',
+  'tedious',
+];
 
 module.exports = {
   mode: process.env.PRODUCTION ? 'production' : 'development',
@@ -11,14 +25,14 @@ module.exports = {
     libraryTarget: 'commonjs2',
     filename: 'index.js',
     devtoolModuleFilenameTemplate: (info) => {
-      const relativePath = path.relative(root, info.absoluteResourcePath)
+      const relativePath = path.relative(root, info.absoluteResourcePath);
       return `webpack://${relativePath}`;
-    }
+    },
   },
   externals: [
     'aws-sdk',
     'electron',
-    { formidable: 'url' }
+    { formidable: 'url' },
   ],
   module: {
     rules: [
@@ -29,7 +43,7 @@ module.exports = {
           {
             loader: 'babel-loader',
             options: {
-              cacheDirectory: true
+              cacheDirectory: true,
             },
           },
         ],
@@ -41,10 +55,13 @@ module.exports = {
   // https://github.com/webpack/webpack/issues/196#issuecomment-620227719
   stats: {
     warningsFilter: [
-      /critical dependency:/i
+      /critical dependency:/i,
     ],
   },
+  plugins: [
+    new IgnorePlugin(new RegExp(`^(${ignoredPackages.join('|')})$`)),
+  ],
   optimization: {
-    nodeEnv: false
-  }
+    nodeEnv: false,
+  },
 };