+++ CHANGELOG.md
@@ -68,9 +68,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - **CUMULUS-2235**
     - Added initial ingest load spec test/utility
   - **CUMULUS-2167**
-    - Added logic to `data-migration2` Lambda for migrating Granule records from DynamoDB to
-      RDS and parse Granule records to
-    store File records in RDS.
+    - Added logic to `data-migration2` Lambda for migrating Granule records from DynamoDB to RDS and parse Granule records to store File records in RDS.
   - **CUMULUS-2367**
     - Added `granules_executions` table to RDS schema to allow for a many-to-many relationship between granules and executions
     - Added `upsertWithExecutionHistory` helper to `@cumulus/db` to allow for upserting a granule record and its corresponding `granules_execution` record
@@ -129,6 +127,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     - Change `PostgresFile` to remove duplicate fields `filename` and `name` and rename `size` to `file_size`.
   - **CUMULUS-2305**
     - Changed `DELETE /pdrs/{pdrname}` API behavior to also delete record from Postgres database.
+  - **CUMULUS-2309**
+    - Changed `DELETE /granules/{granuleName}` API behavior to also delete record from Postgres database.
 
 ### Changed